Fix off-by-one in string slices, could result in panic()
One byte could be left at the end of some template strings,
and if the last character was a grapheme composed of >= 2
bytes, this would panic!().
Also change .chars() for .char_indices() so that we're sure
we don't cut in the middle of a grapheme, even if we're already
careful to only cut after %, !, or other one-byte characters.